Get start repositories and save user if doesan't exist
