+++ src/controllers/repository/getRepoStar.js
@@ -1,30 +1,32 @@
 const fetch = require("node-fetch");
-const repositoriesRepository = require("../../infra/repository/repositories");
+const userRepository = require("../../infra/repository/users");
 const winston = require("../../../config/winston");
+
+const transformer = (repository) => ({
+  name: repository.name,
+  fullName: repository.full_name,
+  descrition: repository.descrition,
+  url: repository.url,
+  html_url: repository.html_url,
+  id: repository.id,
+});
 
 module.exports = async function getRepoStar(req, res) {
   const { username } = req.params;
 
   try {
-    const data = await repositoriesRepository.getRepository({ username });
-    if (data.length > 0) {
-      return res.json({ repo: data, count: data.length });
-    }
+    await userRepository.getUser(username);
 
     const githubStarEndpoint = `https://api.github.com/users/${username}/starred`;
     const repoStarsGithub = await fetch(githubStarEndpoint);
     const repoStarsGithubJson = await repoStarsGithub.json();
-    const repositoriesStars = await repositoriesRepository.saveRepositories(
-      repoStarsGithubJson,
-      username
-    );
 
-    res.json({
-      repo: repositoriesStars,
-      count: repositoriesStars.length,
+    return res.json({
+      repo: repoStarsGithubJson.map(transformer),
+      count: repoStarsGithubJson.length,
     });
   } catch (error) {
     winston.log("error", { message: error });
-    res.status(500).json({ error: error.message });
+    return res.status(500).json({ error: error.message });
   }
 };
