--- src/infra/database/migrate/20200902170541-create-repository.js
@@ -1,39 +1,0 @@
-"use strict";
-module.exports = {
-  up: async (queryInterface, Sequelize) => {
-    await queryInterface.createTable("repositories", {
-      id: {
-        allowNull: false,
-        primaryKey: true,
-        defaultValue: Sequelize.UUIDV1,
-        type: Sequelize.UUID,
-      },
-      name: {
-        type: Sequelize.STRING,
-      },
-      fullName: {
-        type: Sequelize.STRING,
-      },
-      descrition: {
-        type: Sequelize.TEXT,
-      },
-      url: {
-        type: Sequelize.STRING,
-      },
-      htmlUrl: {
-        type: Sequelize.STRING,
-      },
-      createdAt: {
-        allowNull: false,
-        type: Sequelize.DATE,
-      },
-      updatedAt: {
-        allowNull: false,
-        type: Sequelize.DATE,
-      },
-    });
-  },
-  down: async (queryInterface, Sequelize) => {
-    await queryInterface.dropTable("repositories");
-  },
-};
